+++ src/main/java/com/cow/cow_mvc_practice/comment/controller/CommentController.java
@@ -1,7 +1,7 @@
 package com.cow.cow_mvc_practice.comment.controller;
 
-import com.cow.cow_mvc_practice.comment.controller.request.CreateCommentRequest;
-import com.cow.cow_mvc_practice.comment.controller.response.CreateCommentResponse;
+import com.cow.cow_mvc_practice.comment.dto.request.CreateCommentRequest;
+import com.cow.cow_mvc_practice.comment.dto.response.CreateCommentResponse;
 import com.cow.cow_mvc_practice.comment.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class CommentController {
   private final CommentService commentService;
 
-  @PostMapping()
+  @PostMapping("/new")
   public CreateCommentResponse create(@RequestBody final CreateCommentRequest createCommentRequest) {
     return commentService.createComment(createCommentRequest);
   }
