+++ src/main/java/com/cow/cow_mvc_practice/comment/dto/request/CreateCommentRequest.java
@@ -1,4 +1,4 @@
-package com.cow.cow_mvc_practice.comment.controller.request;
+package com.cow.cow_mvc_practice.comment.dto.request;
 
 import com.cow.cow_mvc_practice.comment.entity.Comment;
 import lombok.Builder;
