+++ src/main/java/com/cow/cow_mvc_practice/comment/service/CommentService.java
@@ -1,7 +1,7 @@
 package com.cow.cow_mvc_practice.comment.service;
 
-import com.cow.cow_mvc_practice.comment.controller.request.CreateCommentRequest;
-import com.cow.cow_mvc_practice.comment.controller.response.CreateCommentResponse;
+import com.cow.cow_mvc_practice.comment.dto.request.CreateCommentRequest;
+import com.cow.cow_mvc_practice.comment.dto.response.CreateCommentResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
