+++ src/main/java/com/cow/cow_mvc_practice/member/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.cow.cow_mvc_practice.member.service;
 
-import com.cow.cow_mvc_practice.member.controller.dto.request.CreateMemberRequest;
-import com.cow.cow_mvc_practice.member.controller.dto.request.UpdateMemberRequest;
-import com.cow.cow_mvc_practice.member.controller.dto.response.CreateMemberResponse;
+import com.cow.cow_mvc_practice.member.dto.request.CreateMemberRequest;
+import com.cow.cow_mvc_practice.member.dto.request.UpdateMemberRequest;
+import com.cow.cow_mvc_practice.member.dto.response.CreateMemberResponse;
 import java.util.List;
 
 public interface MemberService {
