+++ src/main/java/com/cow/cow_mvc_practice/member/service/MemberServiceImpl.java
@@ -1,12 +1,12 @@
 package com.cow.cow_mvc_practice.member.service;
 
-import com.cow.cow_mvc_practice.member.controller.dto.request.UpdateMemberRequest;
-import com.cow.cow_mvc_practice.member.controller.dto.response.CreateMemberResponse;
+import com.cow.cow_mvc_practice.member.dto.request.UpdateMemberRequest;
+import com.cow.cow_mvc_practice.member.dto.response.CreateMemberResponse;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import com.cow.cow_mvc_practice.member.controller.dto.request.CreateMemberRequest;
+import com.cow.cow_mvc_practice.member.dto.request.CreateMemberRequest;
 import com.cow.cow_mvc_practice.member.entity.Member;
 import com.cow.cow_mvc_practice.member.repository.MemberJPARepository;
 import jakarta.persistence.EntityNotFoundException;
