+++ src/main/java/com/cow/cow_mvc_practice/post/service/PostService.java
@@ -1,7 +1,7 @@
 package com.cow.cow_mvc_practice.post.service;
 
-import com.cow.cow_mvc_practice.post.controller.dto.request.CreatePostRequest;
-import com.cow.cow_mvc_practice.post.controller.dto.response.CreatePostResponse;
+import com.cow.cow_mvc_practice.post.dto.request.CreatePostRequest;
+import com.cow.cow_mvc_practice.post.dto.response.CreatePostResponse;
 import java.util.List;
 
 public interface PostService {
